Implemented type convertion. Working request with web socket protocol

+++ Sources/Features/TapClient/Sources/TapNLink/Service/AclAPI.swift
@@ -26,10 +26,10 @@ open class AclAPI {
     * LWM2M path: /2/{objectId}/1
 
     * @param objectId input
-    * @return Int api call
+    * @return UInt16 api call
     */
     public func getInstanceIdRequest(
-        objectId: Int
+        objectId: UInt8
     ) -> ApiRequest<Void>{
         var path = "/2/{objectId}/1"
         //var regex = "\{([\S]*)\}"
@@ -50,14 +50,14 @@ open class AclAPI {
         * LWM2M path: /2/{objectId}/1
         
         * @param objectId input
-        * @return Int api call
+        * @return UInt16 api call
         */
         public func getInstanceId(
-            objectId: Int
-        ) throws -> ApiResponse<Int>{
+            objectId: UInt8
+        ) throws -> ApiResponse<UInt16>{
             let request : ApiRequest<Void> = self.getInstanceIdRequest(objectId: objectId)
-            let converter: TapConverterContainer<Int> = self.converterProvider.get(id: "integer_uint16")
-            let response: ApiResponse<Int> = try self.client.execute(request: request, converter: converter)
+            let converter: TapConverterContainer<UInt16> = self.converterProvider.get(id: "integer_uint16")
+            let response: ApiResponse<UInt16> = try self.client.execute(request: request, converter: converter)
 
             return response
         }
@@ -70,7 +70,7 @@ open class AclAPI {
     * @return ListAclEntry api call
     */
     public func getListRequest(
-        objectId: Int
+        objectId: UInt8
     ) -> ApiRequest<Void>{
         var path = "/2/{objectId}/2"
         //var regex = "\{([\S]*)\}"
@@ -94,7 +94,7 @@ open class AclAPI {
         * @return ListAclEntry api call
         */
         public func getList(
-            objectId: Int
+            objectId: UInt8
         ) throws -> ApiResponse<ListAclEntry>{
             let request : ApiRequest<Void> = self.getListRequest(objectId: objectId)
             let converter: TapConverterContainer<ListAclEntry> = self.converterProvider.get(id: "ListAclEntry")
@@ -108,10 +108,10 @@ open class AclAPI {
     * LWM2M path: /2/{objectId}/0
 
     * @param objectId input
-    * @return Int api call
+    * @return UInt16 api call
     */
     public func getObjectIdRequest(
-        objectId: Int
+        objectId: UInt8
     ) -> ApiRequest<Void>{
         var path = "/2/{objectId}/0"
         //var regex = "\{([\S]*)\}"
@@ -132,14 +132,14 @@ open class AclAPI {
         * LWM2M path: /2/{objectId}/0
         
         * @param objectId input
-        * @return Int api call
+        * @return UInt16 api call
         */
         public func getObjectId(
-            objectId: Int
-        ) throws -> ApiResponse<Int>{
+            objectId: UInt8
+        ) throws -> ApiResponse<UInt16>{
             let request : ApiRequest<Void> = self.getObjectIdRequest(objectId: objectId)
-            let converter: TapConverterContainer<Int> = self.converterProvider.get(id: "integer_uint16")
-            let response: ApiResponse<Int> = try self.client.execute(request: request, converter: converter)
+            let converter: TapConverterContainer<UInt16> = self.converterProvider.get(id: "integer_uint16")
+            let response: ApiResponse<UInt16> = try self.client.execute(request: request, converter: converter)
 
             return response
         }
@@ -149,10 +149,10 @@ open class AclAPI {
     * LWM2M path: /2/{objectId}/3
 
     * @param objectId input
-    * @return Int api call
+    * @return UInt16 api call
     */
     public func getOwnerRequest(
-        objectId: Int
+        objectId: UInt8
     ) -> ApiRequest<Void>{
         var path = "/2/{objectId}/3"
         //var regex = "\{([\S]*)\}"
@@ -173,14 +173,14 @@ open class AclAPI {
         * LWM2M path: /2/{objectId}/3
         
         * @param objectId input
-        * @return Int api call
+        * @return UInt16 api call
         */
         public func getOwner(
-            objectId: Int
-        ) throws -> ApiResponse<Int>{
+            objectId: UInt8
+        ) throws -> ApiResponse<UInt16>{
             let request : ApiRequest<Void> = self.getOwnerRequest(objectId: objectId)
-            let converter: TapConverterContainer<Int> = self.converterProvider.get(id: "integer_uint16")
-            let response: ApiResponse<Int> = try self.client.execute(request: request, converter: converter)
+            let converter: TapConverterContainer<UInt16> = self.converterProvider.get(id: "integer_uint16")
+            let response: ApiResponse<UInt16> = try self.client.execute(request: request, converter: converter)
 
             return response
         }
@@ -193,7 +193,7 @@ open class AclAPI {
     * @return  api call
     */
     public func putListRequest(
-        objectId: Int
+        objectId: UInt8
     ) -> ApiRequest<Void>{
         var path = "/2/{objectId}/2"
         //var regex = "\{([\S]*)\}"
@@ -217,7 +217,7 @@ open class AclAPI {
         * @return  api call
         */
         public func putList(
-            objectId: Int
+            objectId: UInt8
         ) throws -> ApiResponse<Void>{
             let request : ApiRequest<Void> = self.putListRequest(objectId: objectId)
             
@@ -237,8 +237,8 @@ open class AclAPI {
     * @return  api call
     */
     public func putOwnerRequest(
-        objectId: Int, ownerId: Int
-    ) -> ApiRequest<Int>{
+        objectId: UInt8, ownerId: UInt32
+    ) -> ApiRequest<UInt32>{
         var path = "/2/{objectId}/3"
         //var regex = "\{([\S]*)\}"
         
@@ -266,9 +266,9 @@ open class AclAPI {
         * @return  api call
         */
         public func putOwner(
-            objectId: Int, ownerId: Int
+            objectId: UInt8, ownerId: UInt32
         ) throws -> ApiResponse<Void>{
-            let request : ApiRequest<Int> = self.putOwnerRequest(objectId: objectId, ownerId: ownerId)
+            let request : ApiRequest<UInt32> = self.putOwnerRequest(objectId: objectId, ownerId: ownerId)
             
             let response: ApiResponse<Void> = try self.client.execute(request: request)
